requirements.txt
anyio==4.9.0
certifi==2025.4.26
decorator==5.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.4
pillow==10.4.0
proglog==0.1.11
python-dotenv==1.1.0
python-telegram-bot==22.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
watchdog==6.0.0
yt-dlp==2025.4.30
